fix projects loading
