+++ .vscode/launch.json
@@ -4,8 +4,7 @@
 // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
 {
 	"version": "0.2.0",
-	"configurations": [
-		{
+	"configurations": [{
 			"name": "Run Extension",
 			"type": "extensionHost",
 			"request": "launch",
@@ -16,7 +15,7 @@
 			"outFiles": [
 				"${workspaceFolder}/out/**/*.js"
 			],
-			"preLaunchTask": "npm: watch"
+			"preLaunchTask": "npm: compile"
 		},
 		{
 			"name": "Extension Tests",
